feat: 로그아웃 확인 메세지

+++ src/hooks/useSignOut.tsx
@@ -3,8 +3,10 @@ import { useNavigate } from 'react-router-dom'
 function useSignOut() {
   const navigate = useNavigate()
   const signOut = () => {
-    localStorage.clear()
-    navigate('/')
+    if (window.confirm('로그아웃하시겠습니까?')) {
+      localStorage.clear()
+      navigate('/')
+    }
   }
 
   return signOut
